+++ modules/bindings/react/connect.js
@@ -1,9 +1,0 @@
-/* @flow */
-import { Component, createElement, PropTypes } from 'react'
-
-import connectFactory from '../connectFactory'
-
-export default connectFactory(Component, createElement, {
-  renderer: PropTypes.object,
-  theme: PropTypes.object
-})