+++ modules/bindings/react/createComponent.js
@@ -1,9 +1,0 @@
-/* @flow */
-import { createElement, PropTypes } from 'react'
-
-import createComponentFactory from '../createComponentFactory'
-
-export default createComponentFactory(createElement, {
-  renderer: PropTypes.object,
-  theme: PropTypes.object
-})