+++ packages/fela-preset-web/index.d.ts
@@ -1,6 +1,0 @@
-declare module "fela-preset-web" {
-  import { TPlugin } from "fela";
-
-  const presets: TPlugin[];
-  export default presets;
-}