+++ scripts/publishPackages.sh
@@ -1,3 +1,0 @@
-for package in packages/*; do
-  cd "$package" && npm publish && cd ../..
-done